separate title and date tests
